feat: Remove unused legacy files from CRM dashboard
This commit removes a significant number of unused files and directories related to a legacy CRM dashboard.

The following changes were made:
- Deleted unused components in `src/components/`, including `dashboard`, `jobs`, `labor`, `layout`, and `common`.
- Deleted unused services in `src/services/`, including `jobService`, `laborService`, and `mockData`.
- Refactored the `Message` type into its own file (`src/types/message.ts`) and deleted the legacy `src/types/job.ts`.
- Removed other unused files like `server.js`, `netlify/functions/test.js`, and the `deploy/` and `dev-dist/` directories.
- Updated `.gitignore` to exclude the `dev-dist/` directory.

These changes significantly clean up the codebase, making it easier to maintain and understand.
