--- netlify/functions/test.js
@@ -1,7 +1,0 @@
-export const handler = async (event, context) => {
-  return {
-    statusCode: 200,
-    headers: { 'Access-Control-Allow-Origin': '*' },
-    body: 'Test function works!'
-  };
-};
