--- src/components/common/Button.tsx
@@ -1,77 +1,0 @@
-import { ReactNode, ButtonHTMLAttributes } from 'react';
-
-type ButtonVariant = 'primary' | 'secondary' | 'success' | 'danger' | 'warning' | 'outline' | 'ghost';
-type ButtonSize = 'sm' | 'md' | 'lg';
-
-interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  children: ReactNode;
-  variant?: ButtonVariant;
-  size?: ButtonSize;
-  isLoading?: boolean;
-  fullWidth?: boolean;
-  leftIcon?: ReactNode;
-  rightIcon?: ReactNode;
-}
-
-const Button = ({
-  children,
-  variant = 'primary',
-  size = 'md',
-  isLoading = false,
-  fullWidth = false,
-  leftIcon,
-  rightIcon,
-  className = '',
-  disabled,
-  ...props
-}: ButtonProps) => {
-  const baseStyles = "inline-flex items-center justify-center font-medium rounded-lg transition-colors focus:outline-none focus:ring-2 focus:ring-offset-2";
-  
-  const variantStyles = {
-    primary: "bg-primary-600 hover:bg-primary-700 text-white focus:ring-primary-500",
-    secondary: "bg-secondary-600 hover:bg-secondary-700 text-white focus:ring-secondary-500",
-    success: "bg-success-500 hover:bg-success-700 text-white focus:ring-success-500",
-    danger: "bg-error-500 hover:bg-error-700 text-white focus:ring-error-500",
-    warning: "bg-warning-500 hover:bg-warning-700 text-white focus:ring-warning-500",
-    outline: "border border-gray-300 dark:border-gray-600 bg-transparent hover:bg-gray-50 dark:hover:bg-gray-800 text-gray-700 dark:text-gray-300 focus:ring-gray-500",
-    ghost: "bg-transparent hover:bg-gray-100 dark:hover:bg-gray-800 text-gray-700 dark:text-gray-300 focus:ring-gray-500",
-  };
-  
-  const sizeStyles = {
-    sm: "text-xs px-3 py-1.5",
-    md: "text-sm px-4 py-2",
-    lg: "text-base px-5 py-2.5",
-  };
-  
-  const widthStyle = fullWidth ? "w-full" : "";
-  
-  const loadingOrDisabled = isLoading || disabled;
-  
-  return (
-    <button
-      className={`
-        ${baseStyles}
-        ${variantStyles[variant]}
-        ${sizeStyles[size]}
-        ${widthStyle}
-        ${loadingOrDisabled ? "opacity-70 cursor-not-allowed" : ""}
-        ${className}
-      `}
-      disabled={loadingOrDisabled}
-      {...props}
-    >
-      {isLoading && (
-        <svg className="animate-spin -ml-1 mr-2 h-4 w-4" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
-          <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-          <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-        </svg>
-      )}
-      
-      {!isLoading && leftIcon && <span className="mr-2">{leftIcon}</span>}
-      {children}
-      {!isLoading && rightIcon && <span className="ml-2">{rightIcon}</span>}
-    </button>
-  );
-};
-
-export default Button;
