--- src/components/common/Input.tsx
@@ -1,104 +1,0 @@
-import { InputHTMLAttributes, forwardRef, useState } from 'react';
-import { Eye, EyeOff } from 'lucide-react';
-
-export interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
-  label?: string;
-  helperText?: string;
-  error?: string;
-  fullWidth?: boolean;
-  leftIcon?: React.ReactNode;
-  rightIcon?: React.ReactNode;
-  showPasswordToggle?: boolean;
-}
-
-const Input = forwardRef<HTMLInputElement, InputProps>(
-  (
-    {
-      label,
-      helperText,
-      error,
-      fullWidth = false,
-      className = '',
-      leftIcon,
-      rightIcon,
-      showPasswordToggle = false,
-      type = 'text',
-      disabled,
-      ...props
-    },
-    ref
-  ) => {
-    const [showPassword, setShowPassword] = useState(false);
-    const inputType = type === 'password' && showPassword ? 'text' : type;
-    
-    const togglePasswordVisibility = () => {
-      setShowPassword(!showPassword);
-    };
-
-    return (
-      <div className={`${fullWidth ? 'w-full' : ''} ${className}`}>
-        {label && (
-          <label 
-            htmlFor={props.id} 
-            className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1"
-          >
-            {label}
-          </label>
-        )}
-        
-        <div className="relative">
-          {leftIcon && (
-            <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-              {leftIcon}
-            </div>
-          )}
-          
-          <input
-            ref={ref}
-            type={inputType}
-            className={`
-              block rounded-md shadow-sm
-              text-gray-900 dark:text-gray-100
-              bg-white dark:bg-gray-800
-              border ${error ? 'border-error-500' : 'border-gray-300 dark:border-gray-600'} 
-              focus:ring-2 focus:ring-primary-500 focus:border-primary-500 dark:focus:ring-primary-700 dark:focus:border-primary-700
-              disabled:bg-gray-100 disabled:dark:bg-gray-900 disabled:cursor-not-allowed disabled:text-gray-500
-              ${leftIcon ? 'pl-10' : 'pl-4'} 
-              ${(rightIcon || (type === 'password' && showPasswordToggle)) ? 'pr-10' : 'pr-4'} 
-              py-2 w-full
-            `}
-            disabled={disabled}
-            {...props}
-          />
-          
-          {type === 'password' && showPasswordToggle && (
-            <button
-              type="button"
-              className="absolute inset-y-0 right-0 pr-3 flex items-center text-gray-500 dark:text-gray-400 focus:outline-none"
-              onClick={togglePasswordVisibility}
-              tabIndex={-1}
-            >
-              {showPassword ? <EyeOff size={18} /> : <Eye size={18} />}
-            </button>
-          )}
-          
-          {rightIcon && !showPasswordToggle && (
-            <div className="absolute inset-y-0 right-0 pr-3 flex items-center pointer-events-none">
-              {rightIcon}
-            </div>
-          )}
-        </div>
-        
-        {(helperText || error) && (
-          <p className={`mt-1 text-sm ${error ? 'text-error-500' : 'text-gray-500 dark:text-gray-400'}`}>
-            {error || helperText}
-          </p>
-        )}
-      </div>
-    );
-  }
-);
-
-Input.displayName = 'Input';
-
-export default Input;
