--- src/components/layout/Header.tsx
@@ -1,37 +1,0 @@
-import { Bell, Sun, Moon } from 'lucide-react';
-import { useTheme } from '../../context/ThemeContext';
-
-interface HeaderProps {
-  title: string;
-}
-
-const Header = ({ title }: HeaderProps) => {
-  const { theme, toggleTheme } = useTheme();
-  
-  return (
-    <header className="bg-white dark:bg-gray-800 shadow-sm border-b border-gray-200 dark:border-gray-700 px-4 py-4 md:py-3 sticky top-0 z-10">
-      <div className="container mx-auto flex justify-between items-center">
-        <h1 className="text-xl md:text-2xl font-bold">{title}</h1>
-        
-        <div className="flex items-center gap-2 md:gap-4">
-          {/* Notifications */}
-          <button className="p-2 rounded-full hover:bg-gray-100 dark:hover:bg-gray-700 relative">
-            <Bell size={20} />
-            <span className="absolute top-1 right-1 w-2 h-2 rounded-full bg-primary-500"></span>
-          </button>
-          
-          {/* Theme Toggle */}
-          <button 
-            className="p-2 rounded-full hover:bg-gray-100 dark:hover:bg-gray-700"
-            onClick={toggleTheme}
-            aria-label={theme === 'dark' ? 'Switch to light mode' : 'Switch to dark mode'}
-          >
-            {theme === 'dark' ? <Sun size={20} /> : <Moon size={20} />}
-          </button>
-        </div>
-      </div>
-    </header>
-  );
-};
-
-export default Header;
