+++ src/types/message.ts
@@ -1,0 +1,8 @@
+export interface Message {
+  id: string;
+  text: string;
+  sender: 'user' | 'ai';
+  timestamp: Date;
+  sessionId?: string;
+  status?: 'sending' | 'sent' | 'delivered' | 'error';
+}
